Masive refactor api
