+++ api/constants.py
@@ -1,0 +1,5 @@
+POKEAPI_URL = "https://pokeapi.co/api/v2"
+
+
+class ERRORS:
+    COULD_NOT_RETRIEVE_DATA = 'Could not retrieve data from the API'
